# specs/001-multi-tenant-ecommerce/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: StormCom Multi-tenant E-commerce API
  version: 1.0.0
  description: |
    RESTful API for the StormCom multi-tenant e-commerce platform.
    
    ## Authentication
    All API requests require authentication via JWT tokens obtained from `/api/auth/login`.
    Include the token in the `Authorization` header: `Bearer <token>`
    
    ## Multi-tenancy
    All requests are scoped to a store via the authenticated user's context.
    Store ID is extracted from the JWT token and auto-injected into queries.
    
    ## Rate Limiting
    Rate limits are enforced per subscription plan:
    - Free: 60 requests/minute
    - Basic: 120 requests/minute
    - Pro: 300 requests/minute
    - Enterprise: 1000 requests/minute
    
    Rate limit headers are included in all API responses (both success and error):
    - `X-RateLimit-Limit`: Maximum requests allowed per 60-second window
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp (seconds) when the window resets
    
    When rate limit is exceeded (HTTP 429), an additional header is included:
    - `Retry-After`: Seconds to wait before retrying the request
    
    ## Pagination
    List endpoints support cursor-based pagination:
    - `page`: Page number (default: 1)
    - `perPage`: Items per page (default: 20, max: 100)
    
    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.

  contact:
...
  license:
...
servers:
...
paths:
...
  /stores:
    get:
      tags: [Stores]
      summary: List stores
      description: List all stores accessible by current user
      operationId: listStores
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, TRIAL, SUSPENDED, CLOSED]
      responses:
        '200':
          description: List of stores
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Store'

    post:
      tags: [Stores]
      summary: Create store
      description: Create new store (requires platform admin role)
      operationId: createStore
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStoreRequest'
      responses:
        '201':
          description: Store created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Store'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from `/auth/login`
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: perPage
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, NOT_FOUND, UNAUTHORIZED, FORBIDDEN, RATE_LIMIT_EXCEEDED, INTERNAL_ERROR]
            message:
              type: string
            details:
              type: object
              description: Additional error context (validation errors, etc.)
    PaginatedResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            page:
              type: integer
            perPage:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer

    # ==================== ENTITY SCHEMAS ====================
...
    Store:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
          nullable: true
        status:
          type: string
          enum: [ACTIVE, TRIAL, TRIAL_EXPIRED, SUSPENDED, CLOSED]
        email:
          type: string
        logoUrl:
          type: string
          nullable: true
        subscription:
          type: object
          properties:
            plan:
              type: string
            status:
              type: string
            trialEndDate:
              type: string
              format: date-time
              nullable: true
        createdAt:
          type: string
          format: date-time
...
    CreateStoreRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 3
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        email:
          type: string
          format: email
        planId:
          type: string
...
  responses:
    BadRequest:
      description: Bad request - Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - BearerAuth: []